fix: prop drilling & styling for deadlines
