+++ src/views/TaxEstimates/TaxEstimates.tsx
@@ -2,15 +2,20 @@ import { useMemo } from 'react'
 import { useTranslation } from 'react-i18next'
 
 import { OnboardingStatus, TaxEstimatesRouteStoreProvider, useTaxEstimatesOnboardingStatus } from '@providers/TaxEstimatesRouteStore/TaxEstimatesRouteStoreProvider'
+import { TaxEstimatesProvider, type TaxEstimatesProviderProps } from '@contexts/TaxEstimatesContext/TaxEstimatesContext'
 import { View } from '@components/View/View'
 import { TaxEstimatesViewContent } from '@views/TaxEstimates/TaxEstimatesViewContent'
 import { TaxEstimatesViewHeader } from '@views/TaxEstimates/TaxEstimatesViewHeader'
 
-export const TaxEstimates = () => {
+export type TaxEstimatesProps = Omit<TaxEstimatesProviderProps, 'children'>
+
+export const TaxEstimates = ({ onTaxBannerReviewClick }: TaxEstimatesProps) => {
   return (
-    <TaxEstimatesRouteStoreProvider>
-      <TaxEstimatesView />
-    </TaxEstimatesRouteStoreProvider>
+    <TaxEstimatesProvider onTaxBannerReviewClick={onTaxBannerReviewClick}>
+      <TaxEstimatesRouteStoreProvider>
+        <TaxEstimatesView />
+      </TaxEstimatesRouteStoreProvider>
+    </TaxEstimatesProvider>
   )
 }
